CNN-PepPred/requirements_CPU.txt
pandas==1.1.3
tensorflow==2.0.0
keras==2.3.1
pathlib
biopython==1.78
logomaker
scikit-learn==0.23.2
seaborn==0.11.0
pillow==8.0.0